+++ src/middleware.rs
@@ -4,5 +4,5 @@
 
 pub use crate::err_handler::ErrorHandlers;
 pub use crate::middleware_from_fn::{from_fn, MiddlewareFn, Next};
-pub use crate::redirect_to_https::redirect_to_https;
+pub use crate::redirect_to_https::RedirectHttps;
 pub use crate::redirect_to_www::redirect_to_www;